requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
zipp==3.13.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
django-bootstrap-v5==1.0.11
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-htmx==1.13.0
django-weasyprint==2.2.0
djangorestframework==3.14.0
fontawesomefree==6.2.1
fonttools==4.38.0
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.4.0
pycparser==2.21
pydyf==0.5.0
pyHanko==0.17.0
pyhanko-certvalidator==0.20.0
pypdf==3.4.0
pyphen==0.13.2
pypng==0.20220715.0
python-barcode==0.14.0
python-bidi==0.4.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==4.4.0
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.14
weasyprint==57.2
webencodings==0.5.1
xhtml2pdf==0.2.9
zopfli==0.2.2

